+++ src/pinetree/gillespie.hpp
@@ -2,7 +2,6 @@
 #define SRC_GILLESPIE_HPP
 
 #include <vector>
-
 #include "reaction.hpp"
 
 class Gillespie {
@@ -19,6 +18,7 @@ class Gillespie {
    * Update propensity of a reaction.
    */
   void UpdatePropensity(Reaction::Ptr reaction);
+  double IndexUpdatePropensity(Reaction::Ptr reaction, int index);
   /**
    * Execute one iteration of the gillespie algorithm.
    */