+++ src/pinetree/polymer.hpp
@@ -26,11 +26,10 @@ class Reaction;
 class MobileElementManager {
  public:
   /**
-   * Only constructor of MobileElementManager
-   *
-   * @param weights Base-pair specific movement weights.
+   * Default constructor needs to be defined here since we have a Python
+   * binding for this class
    */
-  MobileElementManager(const std::vector<double> &weights);
+  MobileElementManager();
   /**
    * Insert an MobileElement-Polymer pair while maintaining order of
    * MobileElements
@@ -79,6 +78,8 @@ class MobileElementManager {
    * @param index Index of MobileElement-Polymer pair
    */
   void UpdatePropensity(int index);
+  void UpdateAllPropensities();
+  void DecrementtRNA(int pol_index);
   /**
    * Getters and setters.
    */
@@ -86,6 +87,8 @@ class MobileElementManager {
   int pol_count() { return pol_count_; }
   int pair_count() const { return polymerases_.size(); }
   int pol_start(int index) const { return polymerases_[index].first->start(); }
+  void set_sequence(const std::string &seq) { seq_ = seq; }
+  void set_weights(const std::vector<double> &weights) { weights_ = weights; }
 
  private:
   /**
@@ -109,8 +112,9 @@ class MobileElementManager {
       std::pair<std::shared_ptr<MobileElement>, std::shared_ptr<Polymer>>>
       polymerases_;
   /**
-   * Base-pair specific movement weights.
+   * Nucleotide sequence of the parent genome.
    */
+  std::string seq_;
   std::vector<double> weights_;
 };
 
@@ -195,6 +199,7 @@ class Polymer : public std::enable_shared_from_this<Polymer> {
    * @param pol polymerase to move
    */
   void Move(int pol_index);
+  void UpdateAllPropensities() { polymerases_.UpdateAllPropensities(); }
   /**
    * Shift mask by 1 base-pair and check for uncovered elements.
    */
@@ -290,13 +295,6 @@ class Polymer : public std::enable_shared_from_this<Polymer> {
    */
   std::map<std::string, int> uncovered_;
   /**
-   * Vector of the same length as this polymer, containing weights for different
-   * positions along the polymer. When a polymerase passes over a given position
-   * in the genome, the weight * speed of polymerase will determine the
-   * propensity for the next movement of that polymerase.
-   */
-  std::vector<double> weights_;
-  /**
    * Finding which binding site (promoter) that the polymerase should bind to.
    *
    * @param pol Pointer to polymerase object
@@ -378,6 +376,7 @@ class Polymer : public std::enable_shared_from_this<Polymer> {
    * @param species_name name of species to uncover
    */
   void LogUncover(const std::string &species_name);
+  void ChoosetRNA(int index);
 };
 
 /**
@@ -402,7 +401,8 @@ class Transcript : public Polymer {
   Transcript(const std::string &name, int start, int stop,
              const std::vector<Interval<BindingSite::Ptr>> &rbs_intervals,
              const std::vector<Interval<ReleaseSite::Ptr>> &stop_site_intervals,
-             const Mask &mask, const std::vector<double> &weights);
+             const Mask &mask, const std::string &seq,
+             const std::vector<double> &weights);
   /**
    * Constructor of transcript used for specifying transcripts without Genome
    *
@@ -439,12 +439,20 @@ class Transcript : public Polymer {
   void AddGene(const std::string &name, int start, int stop, int rbs_start,
                int rbs_stop, double rbs_strength);
   /**
-   * Add transcript weights directly
+   * Add the transcript nucleotide sequence. Normally this is copied from the parent genome,
+   * so this should only be used when adding transcripts to the simulation directly.
    */
-  void AddWeights(const std::vector<double> &transcript_weights);
+  void AddSequence(const std::string &seq);
+  void AddWeights(const std::vector<double> &weights);
+
 
  private:
   std::map<std::string, std::map<std::string, double>> bindings_;
+   /**
+   * Nucleotide sequence of the parent genome.
+   */  
+  std::string seq_;
+  std::vector<double> weights_;
 };
 
 /**
@@ -480,7 +488,16 @@ class Genome : public Polymer {
                int rbs_stop, double rbs_strength);
   void AddRnaseSite(int start, int stop);
   void AddRnaseSite(const std::string &name, int start, int stop, double rnase_degradation_rate);
-  void AddWeights(const std::vector<double> &transcript_weights);
+  /**
+   * Add a nucleotide sequence for this genome. Used to calculate ribosome move propensities
+   * when tRNAs are in the simulation. 
+   */
+  void AddSequence(const std::string &seq);
+  /**
+   * Add translation weights for this genome. Can be used to specify codon-specific translation speeds.
+   * Should not be used if also simulating tRNA dynamics.
+   */
+  void AddWeights(const std::vector<double> &weights);
   const std::map<std::string, std::map<std::string, double>> &bindings();
   const std::map<std::string, double> &rnase_bindings() { return rnase_bindings_; }
   const double &transcript_degradation_rate() {
@@ -510,7 +527,8 @@ class Genome : public Polymer {
   std::vector<Interval<ReleaseSite::Ptr>> transcript_stop_site_intervals_;
   IntervalTree<BindingSite::Ptr> transcript_rbs_;
   IntervalTree<ReleaseSite::Ptr> transcript_stop_sites_;
-  std::vector<double> transcript_weights_;
+  std::string seq_;
+  std::vector<double> weights_;
   std::map<std::string, std::map<std::string, double>> bindings_;
   std::map<std::string, double> rnase_bindings_;
   double transcript_degradation_rate_ = 0.0;