+++ src/pinetree/tracker.hpp
@@ -129,6 +129,16 @@ class SpeciesTracker {
   const std::map<std::string, int> &ribo_per_transcript() {
     return ribo_per_transcript_;
   }
+  void codon_map(const std::map<std::string, std::vector<std::string>> &codon_map) {
+    codon_map_ = codon_map;
+  }
+  const std::map<std::string, std::vector<std::string>> &codon_map() {
+    return codon_map_;
+  }
+  void force_update_all() { force_update_all_ = true; }
+  void unflag_force_update() { force_update_all_ = false; }
+  bool check_force_update() { return force_update_all_; }
+  
   /**
    * Signal to fire when propensity needs to be updated.
    */
@@ -159,6 +169,14 @@ class SpeciesTracker {
    * Species-to-reaction map.
    */
   std::map<std::string, Reaction::VecPtr> species_map_;
+  /**
+   * codon-to-anticodon map.
+   */ 
+  std::map<std::string, std::vector<std::string>> codon_map_;
+  /**
+  * Force gillespie to update all propensities. Currently this should only occur when tRNA pools change.
+  */
+  bool force_update_all_ = false;
 };
 
 #endif  // header guard