+++ tests/models/single_gene.py
@@ -5,8 +5,8 @@ def execute(output):
 
     sim = pt.Model(cell_volume=8e-16)
     sim.seed(34)
-    sim.add_polymerase(name="rnapol", copy_number=1, speed=40, footprint=10)
-    sim.add_ribosome(copy_number=1, speed=30, footprint=10)
+    sim.add_polymerase(name="rnapol", copy_number=1, speed=40.0, footprint=10)
+    sim.add_ribosome(copy_number=1, speed=30.0, footprint=10)
 
     plasmid = pt.Genome(name="T7", length=605)